Function OS_Scaling_onRust::sjf::execute_proccesses

pub fn execute_proccesses(p: &mut Vec<Proccess>)

Executar os processos

Ordena o vetor pelo menor trabalho, e inicializa um loop para executar os processos. Conforme ocorre a execução, o cálculo de Tempo de Espera, Tempo de Turnaround, Tempo médio de espera e Tempo médio de retorno é realizado. Ao fim, imprime todos os processos, juntamente com o resultado dos cálculos e libera a memória executando um pop() de todo o vetor.

p: Endereço mutável de um vetor de Processos

medium_return_time: Tempo Médio de Retorno

medium_await_time: Tempo Médio de Espera

_popped_proc: Processo que foi removido #variável não utilizada#

total_time: Tempo total decorrido durante a execução em u.t.